Warn user if both size and the time interval options are used
The documentation says (in the section for (max,min)size) that
the size option is mutually exclusive with the time interval option
and it says:

```
The related size option
is similar except that it is mutually exclusive with the  time
interval  options, *and it causes log files to be rotated with-
out regard for the last rotation time.*
```

(emphasis mine).

The emphasised bit of the sentence is not necessarily true - it is
true _if the size option is specified after the time option_. If
the time option is specified after the size option, the logrotate
will rotate the files without regard for the size. In other words -
the last option specified in the config file wins.

This behaviour is dangerous, because the subtle change in the order
of the configuration options changes the behaviour in the unexpected
way.

We can't treat this behaviour as an error, since the chances are that
there are many logrotate config files deployed in the way that it works
to the satisfactory degree, so that would be a breaking change. This
patch makes logrotate to issue a warning if logrotate run in verbose mode.

Closes #192